refactor: Reorganize src files into folders
